--- kotlin_feature_identity/build.gradle.kts
@@ -1,79 +1,0 @@
-import com.vanniktech.maven.publish.SonatypeHost
-
-plugins {
-    alias(libs.plugins.android.library)
-    alias(libs.plugins.jetbrains.kotlin.android)
-    id("maven-publish")
-
-    id("com.vanniktech.maven.publish") version "0.29.0"
-}
-
-android {
-    namespace = "com.fadlurahmanfdev.kotlin_feature_identity"
-    compileSdk = 34
-
-    defaultConfig {
-        minSdk = 21
-
-        testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
-        consumerProguardFiles("consumer-rules.pro")
-    }
-
-    buildTypes {
-        release {
-            isMinifyEnabled = false
-            proguardFiles(
-                getDefaultProguardFile("proguard-android-optimize.txt"),
-                "proguard-rules.pro"
-            )
-        }
-    }
-    compileOptions {
-        sourceCompatibility = JavaVersion.VERSION_1_8
-        targetCompatibility = JavaVersion.VERSION_1_8
-    }
-    kotlinOptions {
-        jvmTarget = "1.8"
-    }
-}
-
-dependencies {
-    testImplementation(libs.junit)
-    androidTestImplementation(libs.androidx.junit)
-    androidTestImplementation(libs.androidx.espresso.core)
-
-    api("androidx.biometric:biometric:1.1.0")
-}
-
-mavenPublishing {
-    publishToMavenCentral(SonatypeHost.CENTRAL_PORTAL)
-    signAllPublications()
-
-    coordinates("com.fadlurahmanfdev", "kotlin_feature_identity", "0.0.1-beta")
-
-    pom {
-        name.set("Kotlin Library Feature Identity")
-        description.set("A comprehensive identity library for Android, designed to provide seamless integration with features like biometric authentication, passkeys, credential management, and more, based on Android's identity framework.")
-        inceptionYear.set("2024")
-        url.set("https://github.com/fadlurahmanfdev/kotlin_feature_identity/")
-        licenses {
-            license {
-                name.set("The Apache License, Version 2.0")
-                url.set("http://www.apache.org/licenses/LICENSE-2.0.txt")
-                distribution.set("http://www.apache.org/licenses/LICENSE-2.0.txt")
-            }
-        }
-        developers {
-            developer {
-                id.set("fadlurahmanfdev")
-                name.set("Taufik Fadlurahman Fajari")
-                url.set("https://github.com/fadlurahmanfdev/")
-            }
-        }
-        scm {
-            url.set("https://github.com/fadlurahmanfdev/kotlin_feature_identity/")
-            connection.set("scm:git:git://github.com/fadlurahmanfdev/kotlin_feature_identity.git")
-            developerConnection.set("scm:git:ssh://git@github.com/fadlurahmanfdev/kotlin_feature_identity.git")
-        }
-    }
-}
