--- kotlin_feature_identity/src/main/java/com/fadlurahmanfdev/kotlin_feature_identity/core/enums/CheckAuthenticationStatusType.kt
@@ -1,7 +1,0 @@
-package com.fadlurahmanfdev.kotlin_feature_identity.core.enums
-
-enum class CheckAuthenticationStatusType {
-    BIOMETRIC_WEAK,
-    BIOMETRIC_STRONG,
-    DEVICE_CREDENTIAL,
-}
