--- kotlin_feature_identity/src/main/java/com/fadlurahmanfdev/kotlin_feature_identity/core/enums/FeatureAuthenticatorType.kt
@@ -1,13 +1,0 @@
-package com.fadlurahmanfdev.kotlin_feature_identity.core.enums
-
-enum class FeatureAuthenticatorType {
-    /**
-     * Represents biometric authentication using weak biometric
-     */
-    BIOMETRIC,
-
-    /**
-     * Represents biometric authentication using device credential (PIN, PASSWORD, or PATTERN).
-     */
-    DEVICE_CREDENTIAL
-}
