--- kotlin_feature_identity/src/main/java/com/fadlurahmanfdev/kotlin_feature_identity/core/exception/FeatureIdentityException.kt
@@ -1,6 +1,0 @@
-package com.fadlurahmanfdev.kotlin_feature_identity.core.exception
-
-data class FeatureIdentityException(
-    val code: String,
-    override val message: String? = null,
-) : Throwable(message = message)
