+++ mark_biometric/proguard-rules.pro
