+++ mark_biometric/src/main/AndroidManifest.xml
