+++ mark_biometric/src/main/java/com/fadlurahmanfdev/mark_biometric/core/constant/ErrorConstant.kt
@@ -1,0 +1,13 @@
+package com.fadlurahmanfdev.mark_biometric.core.constant
+
+object ErrorConstant {
+    const val OS_NOT_SUPPORTED = "OS_NOT_SUPPORTED"
+    const val CIPHER_MISSING = "CIPHER_MISSING"
+    const val UNABLE_FETCH_GET_SECRET_KEY = "UNABLE_FETCH_GET_SECRET_KEY"
+    const val SECRET_KEY_MISSING = "SECRET_KEY_MISSING"
+    const val UNABLE_SECURE_AUTHENTICATE = "UNABLE_SECURE_AUTHENTICATE"
+    const val KEY_PERMANENTLY_INVALIDATED = "KEY_PERMANENTLY_INVALIDATED"
+    const val UNABLE_TO_DELETE_SECRET_KEY = "UNABLE_TO_DELETE_SECRET_KEY"
+    const val UNABLE_TO_DETECT_BIOMETRIC_CHANGE = "UNABLE_TO_DETECT_BIOMETRIC_CHANGE"
+    const val BAD_PADDING = "BAD_PADDING"
+}
