+++ mark_biometric/src/main/java/com/fadlurahmanfdev/mark_biometric/core/enums/FeatureAuthenticatorType.kt
@@ -1,0 +1,13 @@
+package com.fadlurahmanfdev.mark_biometric.core.enums
+
+enum class FeatureAuthenticatorType {
+    /**
+     * Represents biometric authentication using weak biometric
+     */
+    BIOMETRIC,
+
+    /**
+     * Represents biometric authentication using device credential (PIN, PASSWORD, or PATTERN).
+     */
+    DEVICE_CREDENTIAL
+}
